[FIX] l10n_ar_ux: monkey patch post_load
this way the monkey patch only is loaded if the module is installed.

X-original-commit: ccf2b07e6eed8e1e2e0d4cc90dbb7b6a17d8529c
